LPS-111990 add languague properties for fieldset and refactor saveAsFieldSet function
