+++ modules/apps/app-builder/app-builder-lang/src/main/resources/content/Language.properties
@@ -33,6 +33,7 @@ edit-table-view=Edit Table View
 entry-added=Entry Added
 entry-deleted=Entry Deleted
 entry-updated=Entry Updated
+fieldset-saved=Fieldset saved
 file-preview=File Preview
 filter-by-author=Filter by Author
 filter-by-deployment-type=Filter by Deployment Type
@@ -63,6 +64,7 @@ place-it-in-the=Place it in the
 previous-entry=Previous Entry
 product-menu=Product Menu
 retrieving-all-objects=Retrieving all objects...
+save-as-fieldset=Save As Fieldset
 select-a-form-view=Select a Form View
 select-a-table-view=Select a Table View
 select-object=Select Object
